Move client rest api bits to pbench namespace
- Move the agent rest api bits to the pbench namespace.
- Remove py3 testing envrionment.
- Split out tests into functional and unit.
- Update tox.ini
- Updated module imports.
