+++ .travis.yml
@@ -9,7 +9,6 @@ env:
   - TOXENV=server
   - TOXENV=pytest_agent
   - TOXENV=pytest_server
-  - TOXENV=py3
 before_install:
  - sudo apt-get update
 install:
