--- pbench/test/test_pbench.py
@@ -1,2 +1,0 @@
-def test_placeholder():
-    pass
