+++ pbench/test/unit/agent/common.py
@@ -3,10 +3,10 @@ import tempfile
 
 FAKE_TIME = datetime.datetime(2020, 2, 2, 22, 0, 0)
 target_dir = tempfile.mkdtemp()
-tarball = "agent/lib/pbench/tests/fixtures/copy_result_tb/log.tar.xz"
+tarball = "pbench/test/unit/agent/fixtures/copy_result_tb/log.tar.xz"
 bad_tarball = "nothing.tar.xz"
 
-MRT_DIR = "agent/lib/pbench/tests/fixtures/make_result_tb"
+MRT_DIR = "pbench/test/unit/agent/fixtures/make_result_tb"
 
 
 class MockDatetime(datetime.datetime):
@@ -16,4 +16,4 @@ class MockDatetime(datetime.datetime):
 
 
 def mock_agent_config(*args, **kwargs):
-    return {"pbench_run": "agent/lib/pbench/tests/fixtures/move_results"}
+    return {"pbench_run": "pbench/test/unit/agent//fixtures/move_results"}
