+++ pbench/test/unit/agent/fixtures/config/pbench-agent.cfg
@@ -1,5 +1,5 @@
 [DEFAULT]
-pbench_install_dir = ../../../../..
+pbench_install_dir = ../../../../../../agent/
 pbench_results_redirector = pbench.results.example.com
 pbench_web_server = pbench.example.com
