+++ tox.ini
@@ -3,7 +3,7 @@ minversion = 3.5.0
 skipdist = True
 skip_missing_interpreters = True
 isolated_build = True
-envlist = lint,server,agent,pytest_server,pytest_agent,py3
+envlist = lint,server,agent,pytest_server,pytest_agent
 
 [testenv]
 usedevelop = True
@@ -28,13 +28,6 @@ commands_pre =
 commands =
   bash -c "./server/bin/unittests"
 
-[testenv:py3]
-usedevelop = true
-commands_pre = 
-  find {toxinidir} -type f -not -path '{toxinidir}/.tox/*' -path '*/__pycache__/*' -name '*.py[c|o]' -delete
-commands =
-  pytest -s ./pbench/test
-
 [testenv:agent]
 usedevelop = true
 commands_pre = 
@@ -53,12 +46,11 @@ commands =
 
 [testenv:pytest_agent]
 setenv =
-  PYTHONPATH = {envdir}{:}{toxinidir}{:}{toxinidir}/agent/lib/
-  _PBENCH_AGENT_CONFIG={toxinidir}/agent/lib/pbench/tests/fixtures/config/pbench-agent.cfg
+  _PBENCH_AGENT_CONFIG={toxinidir}/pbench/test/unit/agent/fixtures/config/pbench-agent.cfg
 commands_pre =
   find {toxinidir} -type f -not -path '{toxinidir}/.tox/*' -path '*/__pycache__/*' -name '*.py[c|o]' -delete
 commands =
-  pytest ./agent/lib/pbench/tests
+  pytest ./pbench/test/unit/agent
 
 [testenv:lint]
 description = Runs all linting tasks
